requirements.txt
dash>=2.14.0
plotly>=5.17.0
pandas>=2.0.0
numpy>=1.25.0
scipy>=1.11.0
scikit-learn>=1.3.0
ccxt>=4.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
redis>=5.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0